requirements.txt
tensorflow
scikit-learn
opencv-python-headless
numpy
Pillow
matplotlib
seaborn
tqdm
albumentations
joblib
streamlit